refactor(login/login-email-confirmation): revise domain and useCountdown
